lab3/exercice1/requirements.txt
apache-libcloud
paramiko
pyyaml